wireguard: Add fwmark and rules to ensure tunnel only runs over wan.
When no default route was present in the table main, but only in table
fff because of disconnected WAN, wireguard was digging it's tunnel
through fff-network.

Depending on the metrics, the gateway therefore used the tunnel as
default-route which let the tunnel collapse.

By introducing fwmark and routing it seperately, this can be avoided.

Signed-off-by: Fabian Bläse <fabian@blaese.de>