gre: Disable Host Dependency
A tunnel cannot be set up, if the peer configured is unreachable for the
route when the tunnel is set up, because adding host dependency fails.

As already stated when this dependency was disabled for wireguard, it
seems to be somehow broken for IPv6 peers as well.

Because it is not necessary for us anyway, host dependency is disabled
for gre with this patch.

Signed-off-by: Fabian Bläse <fabian@blaese.de>